Change set-up of Tooltip server render test
